handling backstack

+++ app/src/main/java/ru/mvrlrd/playlistmaker/database/data/FavoriteDao.kt
@@ -11,11 +11,11 @@ interface FavoriteDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertTrack(trackEntity: TrackEntity)
     @Query("DELETE FROM favorite_tracks WHERE id =:id")
-    suspend fun deleteTrack(id: Int)
+    suspend fun deleteTrack(id: Long)
     @Query("SELECT * FROM favorite_tracks ORDER by date DESC")
     suspend fun getFavoriteTracks(): List<TrackEntity>
     @Query("SELECT id FROM favorite_tracks")
-    suspend fun getFavoriteTrackIds(): List<Int>
+    suspend fun getFavoriteTrackIds(): List<Long>
     @Query("DELETE FROM favorite_tracks")
     suspend fun clearFavorites()
