+++ app/src/main/java/ru/mvrlrd/playlistmaker/db_playlist/entities/PlaylistEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "playlist_table")
 class PlaylistEntity(
     @PrimaryKey(autoGenerate = true)
-    val playlistId: Int = -1,
+    val playlistId: Long = -1,
     val name: String,
     val description: String,
     val playlistImagePath: String,
