+++ app/src/main/java/ru/mvrlrd/playlistmaker/db_playlist/entities/PlaylistSongCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["playlistId", "songId"])
 data class PlaylistSongCrossRef(
-    val playlistId: Int,
-    val songId: Int
+    val playlistId: Long,
+    val songId: Long
 )
