+++ app/src/main/java/ru/mvrlrd/playlistmaker/mediateka/playlists/addplaylist/domain/PlaylistForAdapter.kt
@@ -1,7 +1,7 @@
 package ru.mvrlrd.playlistmaker.mediateka.playlists.addplaylist.domain
 
 data class PlaylistForAdapter(
-    val playlistId: Int?=null,
+    val playlistId: Long? = null,
     val name: String,
     val description: String,
     val playlistImagePath: String,
