+++ app/src/main/java/ru/mvrlrd/playlistmaker/search/data/model/TrackDto.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 
 data class TrackDto(
-    val trackId: Int,
+    val trackId: Long,
     val trackName: String,
     val artistName: String,
     @SerializedName("trackTimeMillis") val trackTime: String,
@@ -26,7 +26,7 @@ data class TrackDto(
     }
 
     override fun hashCode(): Int {
-        return trackId
+        return trackId.toInt()
     }
     fun getCoverArtwork() = image.replaceAfterLast('/',"512x512bb.jpg")
 }
