+++ app/src/main/java/ru/mvrlrd/playlistmaker/search/domain/TracksInteractor.kt
@@ -8,6 +8,6 @@ interface TracksInteractor {
     fun addTrackToHistory(adapterTrack: AdapterTrack)
     fun clearHistory()
     suspend fun getHistory(): Flow<List<AdapterTrack>>
-    suspend fun getFavIds():Flow<List<Int>>
+    suspend fun getFavIds():Flow<List<Long>>
 }
